Move react-helpers/custom-component.ts to react-helpers/CustomComponentIFrame/path.ts
