+++ packages/stlite-kernel/src/react-helpers/CustomComponentIFrame/path.ts
